+++ test/source/dvd.h
@@ -134,8 +134,8 @@ struct Patch
 
     /* file info */
     u8 drv;
-    DWORD start_cluster;
-    DWORD cur_cluster;
+    u64 start_cluster;
+    u64 cur_cluster;
     u32 file_offset;
 };
